chore(deps): break PilotRepository ↔ career import cycle
Extract PilotErrorCode + IPilotOperationResult into a leaf
PilotRepository.types.ts module so the repository class and its
career-operation helpers no longer import from each other.
PilotRepository.ts re-exports both names so external consumers
(api/pilots, PilotService, tests, the pilots barrel index) keep their
existing import paths working.

Cycle removed: pilots/PilotRepository.ts ↔ pilots/PilotRepository.career.ts
(madge --circular: 11 → 10)

No behavior change.
